Use api package for move AuthenticationDescriptor classes
